Fix crashes with CGSPrivate

+++ Quicksilver/Code-External/CGSPrivate.h
@@ -29,9 +29,9 @@
  */
 
 // Internal CoreGraphics typedefs
-typedef NSInteger CGSConnection;
-typedef NSInteger CGSWindow;
-typedef NSInteger CGSValue;
+typedef int CGSConnection;
+typedef int CGSWindow;
+typedef int CGSValue;
 
 //// CONSTANTS ////
 
@@ -209,7 +209,7 @@ struct CGPointWarp {
 };
 
 extern CGError CGSSetWindowWarp(const CGSConnection cid, const CGSWindow wid, uint32_t w, uint32_t h, CGPointWarp *mesh);
-extern OSStatus CGSSetWindowListBrightness(const CGSConnection cid, CGSWindow *wids, CGFloat *brightness, NSInteger count);
+extern OSStatus CGSSetWindowListBrightness(const CGSConnection cid, CGSWindow *wids, float *brightness, int count);
 
 // Global Hotkey
 typedef enum {
